Completed Lens System & implemented a DatabaseCore
Lens system has been implemented and is finished - Still more
improvements to make however it is entirely functional and additions are
minor.

Added a DatabaseCore:
- The DBCore is designed to allow for easy implementation of new
databases.
- It uses a CallbackHandler to assign methods to be invoked by the
system in order to parse the query results in specific ways. For
example; You can query a database and invoke a method to build specific
objects that are returned to the system and passed back to the caller
(In either a list or single object depending on the called function)